tc: Extract common code for getting a symbol
For both qdiscs and filters, the code first looks for symbol whose name is
derived from kind, in a DSO whose name is derived from kind (falling back
to current binary). Extract this common code into a helper so that it does
not have to be repeated in various functions across TC.

Note that previously, m_pedit and m_ematch did not specify full path to the
TC plugin. I believe this is not intentional, and have migrated that code
over to the new get_symbol() as well.

m_ipt uses a more complex pattern of dlopens, which was left as is.

Signed-off-by: Petr Machata <petrm@mellanox.com>