tc: q_fifo: Support new-style dumps
The only interface to configure FIFO is currently a four-byte datum with the
FIFO length. In order to permit FIFO extensions, kernel now accepts additional
set of netlink attributes after this field. Dumping was similarly
extended--attributes are attached after the four-byte field.

Adapt q_fifo to use this new mode for pfifo and bfifo qdiscs.

Signed-off-by: Petr Machata <petrm@mellanox.com>