lib: json_print: Add print_color_fmt()
A number of sbuf users has the following form:

	SPRINT_BUF(b1);
	snprintf(b1, sizeof(b1), "%#.2hhx", field);
	print_string(PRINT_ANY, "field", "field %s ", b1);

With this being a somewhat pervasive pattern, perhaps we should have a
helper that encapsulates the logic. In this patch, introduce a new print
helper, print_color_fmt() / print_fmt(), which produces a temporary string
according to a given format and given arguments, and then uses the
temporary to emit a string field. The above then becomes:

	print_fmt(PRINT_ANY, "field", "field %s ", "%#.2hhx", mode);

Signed-off-by: Petr Machata <petrm@nvidia.com>