lib: utils: Convert ppp_proto_n2a() to use sbuf's
Instead of relying on a passed-in statically-sized buffer, have the
function take an sbuf.

Since this function is just a wrapper around proto_n2a(), which still
relies on pre-sized buffers, the function temporarily needs to become
an adaptor. A follow-up fatch converts the latter function as well, at
which point the adaptation disappears.

Signed-off-by: Petr Machata <petrm@nvidia.com>