xxx
A lot of macro magic to get a fairly meager benefit. The ability to ad hoc
declare/define arrays with a given formatted content is pretty cool. But
it's not useful that often, most of the SPRINT_BUF uses are things like
ll_addr_a2n() where the buffer is used from inside a called function.