lib: Change behavior of parse_one_of()
The function parse_one_of() currently uses matches() for string comparison
under the hood. Extending matches()-based parsers is tricky, because newly
added matches might change the way strings are parsed, if the newly-added
string shares a prefix with a string that is matched later in the code.

Therefore in this patch, change parse_one_of() to matches(), but with a
twist that prefix matches yield a warning. This will not disturb standard
output or the overall behavior, but will make it obvious that the usage is
discouraged and prompt users to update their scripts and habits.

Signed-off-by: Petr Machata <petrm@nvidia.com>